接入微信支付 修改获取价格参数为int

+++ app/src/main/java/com/heixiu/errand/net/ApiService.java
@@ -141,7 +141,7 @@ public interface ApiService {
             @Query("receiveNum") String receiveNum,
             @Query("courierNum") String courierNum,
             @Query("supportPrice") int supportPrice,
-            @Query("payment") double payment,
+            @Query("payment") int payment,
             @Query("originsLatitude") String originsLatitude,
             @Query("originsLongitude") String originsLongitude,
             @Query("destinationsLatitude") String destinationsLatitude,
